Added Safety
Added more safety checks, as some states can have null nodes still. In the case off null-nodes, it'll still run, but will generate errors in the log. This will prevent odd crashing, but will generate the appropriate warning to make sure nodes are set properly.

Future goal is to see if we can set default values for Ustructs with UObjects.