Got Compile toolbar working
Compile toolbar is currently there for the GraphMode, but not BlueprintMode. Also of note
that a line of code, specifically "BP->RegisterToolbarTab(InTabManager.ToSharedRef());" is not being used for this currently. It only brings up a window with extra tabs that aren't used, and has done unrecognized tab. For now, we'll leave it out, as the toolbar is working, i.e. appearing.