Working on linking EditActions for the Graph
Currently nodes can be added, renamed, but selection is an issue for the inspector. This is
due to the fact that the graph doesn't actually handle node selection, so we'll need to
implement node selection tracking later. Also need put back the initial text for the
state nodes, as well as fix the new state node naming scheme.