SM Blueprints can be compiled
Currently SM blueprints can be compiled. This required a bit of a change for the work flow.

SM's made in blueprints can be instanced in default properties, however the graph made in the editor cannot be made available (either a limitation or requires extra work that we have no documentation on). So currently default values for SM's are used as overrides or addictions to the graph. The graph is moved into the SM upon initialization in a component or called directly, and can be used.