Implemented Inherited Object Model
State Machine Blueprints need to utilize an object inheritance model, in a way that it can override
parts of an object defined in the parent. In this case the statemachine types used for subgraphs. This can't be safely done by just swapping out another StateMachine as many assumptions are made about what statemachine the parent is. Thus we've implemented a way to utilize an archetype object saved in the SMBP Generated Classes and in Override EdStateGraphs so that we can check if values have been changed, and if so update them as necessary.

TODO:
- Implement recursive structure copying.