Updated PatrolPath + ElevatorSystem
The PatrolPath actor has been updated a bit to have a cleaner API. Before the states used took a
path in as a parameter, but since it isn't really well-defined to use states between different
paths, we changed it so a state also contains a path to work with.

The elevator system was also updated to use a patrolpathstate instead of maintaining its own
patrol-like state. This simplifies the system a bit, and utilizes the editor rendering for paths a bit better.

TODO: Make patrol paths also work between both rigid (straight-line) paths and curved paths, as the elevator may want to use the path exactly.